% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{bootstrap_simulations}
\alias{bootstrap_simulations}
\title{Title}
\usage{
bootstrap_simulations(
  uids,
  ind,
  gformula,
  splits,
  newdata,
  interventions,
  n_sims,
  n_samples,
  start_time,
  stop_time,
  bound_sims,
  last_only,
  prediction
)
}
\arguments{
\item{prediction}{}
}
\description{
Title
}
