% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_bounded_continuous}
\alias{fit_bounded_continuous}
\title{Fit Bounded Normal Model on Covariate}
\usage{
fit_bounded_continuous(formula, link = NULL, data, control = NULL)
}
\arguments{
\item{control}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate at index \eqn{j}.
}\if{html}{\out{</div>}}
}
\description{
This internal function models a covariate using a "bounded normal" distribution
by first standardizing the covariate values to the range \link{0, 1}, noninclusive,
then fitting a generalized linear model (GLM) under the Gaussian family function.
}
\keyword{internal}
