% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_compevent_model}
\alias{fit_compevent_model}
\title{Fit Competing Event Model}
\usage{
fit_compevent_model(model, data, use_gam = FALSE)
}
\arguments{
\item{model}{Model statement for competing event variable.}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula),
typically the environment from which glm is called.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                  Fitted model for the competing event variable.
}\if{html}{\out{</div>}}
}
\description{
This internal function fits a generalized linear model (GLM) for the competing event variable using the observed data.
}
\keyword{internal}
