% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_covariate_model}
\alias{fit_covariate_model}
\title{Fit Covariate Models}
\usage{
fit_covariate_model(models, data, time = "time", start_time = 0, gam = FALSE)
}
\arguments{
\item{data}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           A list of fitted models, one for each covariate in \code{covnames}.
}\if{html}{\out{</div>}}
}
\description{
This internal function fits a model for each covariate using the observed data.
}
\keyword{internal}
