% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit GLM on Covariate}
\usage{
fit_glm(formula, family, link = NULL, data, control = NULL)
}
\arguments{
\item{control}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate.
}\if{html}{\out{</div>}}
}
\description{
This internal function fits a generalized linear model (GLM) for a single covariate using the observed data.
}
\keyword{internal}
