% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_hurdle}
\alias{fit_hurdle}
\title{Fit Hurdle Model for Counts on Covariate}
\usage{
fit_hurdle(formula, family, link, data, control = NULL)
}
\arguments{
\item{control}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate at index \eqn{j}.
}\if{html}{\out{</div>}}
}
\description{
This internal function models a covariate using a hurdle model, i.e. a
two-component model in which zeros are modeled seperately from a truncated
distribution for positive counts. The count model can either be a truncated
Poisson or negative binomial model. The zero or "hurdle" component is
modeled with a binomial GLM (e.g. logistic regression).
}
\keyword{internal}
