% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_multinomial}
\alias{fit_multinomial}
\title{Fit Multinomial Model on Covariate}
\usage{
fit_multinomial(formula, data, control = NULL)
}
\arguments{
\item{control}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate at index \eqn{j}.
}\if{html}{\out{</div>}}
}
\description{
This internal function fits a multinomial regression model for a categorical covariate using the observed data.
}
\keyword{internal}
