% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_outcome_model}
\alias{fit_outcome_model}
\title{Fit Outcome Model}
\usage{
fit_outcome_model(
  model,
  data,
  time = NULL,
  stop_time = NULL,
  survival = FALSE,
  gam = FALSE
)
}
\arguments{
\item{model}{Model statement for the outcome variable.}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula),
typically the environment from which glm is called.}

\item{time}{Time variable.}

\item{EOF}{Only use end of follow up time point when fitting outcome model}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          Fitted model for the outcome variable.
}\if{html}{\out{</div>}}
}
\description{
This internal function fits a generalized linear model (GLM) for the outcome variable using the observed data.
}
\keyword{internal}
