% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_trunc_normal}
\alias{fit_trunc_normal}
\title{Fit Truncated Normal Model on Covariate}
\usage{
fit_trunc_normal(
  formula,
  link = NULL,
  data,
  control = NULL,
  direction = NULL,
  point = NULL
)
}
\arguments{
\item{point}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate at index \eqn{j}.
}\if{html}{\out{</div>}}
}
\description{
This internal function models a covariate using a normal distribution truncated on one
side at a user-specified cutoff.
}
\keyword{internal}
