% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_zeroinfl_count}
\alias{fit_zeroinfl_count}
\title{Fit Zero-inflated Model for Counts on Covariate}
\usage{
fit_zeroinfl_count(formula, family, link, data, control = NULL)
}
\arguments{
\item{control}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate at index \eqn{j}.
}\if{html}{\out{</div>}}
}
\description{
This internal function models a covariate using a zero-inflated count model,
i.e. a two-component mixture model in which observations are assumed to
arise from a mixture distribution composed of a 0/1 distribution such as the
bernoulli/binomial and a proper count distribution. It differs from a "hurdle"
model in that there are two sources of zeros: zeros may come from the 0/1
distribution or from the proper count distribution.
}
\keyword{internal}
