% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_zeroinfl_normal}
\alias{fit_zeroinfl_normal}
\title{Fit Zero-Inflated Normal Model on Covariate}
\usage{
fit_zeroinfl_normal(formula, link = NULL, data, control = NULL)
}
\arguments{
\item{control}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Fitted model for the covariate at index \eqn{j}.
}\if{html}{\out{</div>}}
}
\description{
This internal function models a zero-inflated normal distribution through the combined
use of a generalized linear model (GLM) fit on a zero vs. non-zero indicator
and a GLM fit on all non-zero values.
}
\keyword{internal}
