% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula-function.R
\name{gformula}
\alias{gformula}
\title{Causal inference by g-formula}
\usage{
gformula(formula, data, bootstrap = 30, state = 33, verbose = F)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the
model to be fitted. The exposure of interest should be plugged-in as the
first covariate at the right-hand size of the formula.}

\item{data}{A data frame containing the variables in the model.}

\item{bootstrap}{An integer determining how many times this procedure
being repeated by resampling with replacement.}

\item{state}{An integer to set random seed for reproducible results.}

\item{verbose}{A logical determining whether a progress bar is shown.}
}
\value{
output A list containing the formula, exposure of interest, marginal
effect, 95\% confidence interval (CI), significance by p-value obtained from
the CI (https://doi.org/10.1136/bmj.d2304), data, bootstrapping times,
random seed, and index for each bootstrap set.
}
\description{
This function conduct causal inference by implementing g-formula. Please
read (https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book/)
before applying this test.
}
\examples{

# Load example data for formula and data
input=input_example()
formula=input$formula
data=input$data

# Conduct g-formula
gformula(formula,data)
}
\keyword{g-formula}
