% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmethods-package.R
\docType{package}
\name{gmethods-package}
\alias{gmethods}
\alias{gmethods-package}
\title{gmethods: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://christopherbboyer.com/gmethods/}
}

}
\author{
\strong{Maintainer}: Christopher Boyer \email{christopher.b.boyer@gmail.com} (\href{https://orcid.org/0000-0002-1101-2045}{ORCID})

}
\keyword{internal}
