% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_example-function.R
\name{input_example}
\alias{input_example}
\title{Make an input example for gmethods package}
\usage{
input_example()
}
\value{
output A list of a formula and a data frame with dimension of 1629
rows and 10 columns.
}
\description{
This function load a causal model as formula object and a data frame of
nhefs table, a cleaned version of NHEFS data. In Causal Inference: What If
book by Hernán and Robins, four NHEFS datasets are described in this book
(https://www.hsph.harvard.edu/miguel-hernan/causal-inference-book/). The
remaining datasets from that book could be retrieve from cidata R package
(https://github.com/malcolmbarrett/cidata) for more information.
}
\examples{

# Load example data for formula and data
input=input_example()
formula=input$formula
data=input$data
}
\keyword{data}
\keyword{example}
