% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{run_gformula}
\alias{run_gformula}
\title{Title}
\usage{
run_gformula(
  gformula,
  newdata = NULL,
  interventions = NULL,
  n_sims = 1,
  n_samples = NULL,
  n_boots = NULL,
  start_time = NULL,
  stop_time = NULL,
  bound_sims = FALSE,
  return_sims = FALSE,
  return_covs = TRUE,
  last_only = TRUE,
  natural_course = TRUE,
  conf_level = 0.95,
  conf_type = "norm"
)
}
\arguments{
\item{conf_type}{}
}
\description{
Title
}
