% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{simulate.gformula}
\alias{simulate.gformula}
\title{Run Monte Carlo simulation to estimate population means under time-varying
interventions using (parametric) g-formula}
\usage{
\method{simulate}{gformula}(
  gformula,
  newdata = NULL,
  interventions = NULL,
  reference = "Natural course",
  n_samples = NULL,
  n_boots = NULL,
  start_time = NULL,
  stop_time = NULL,
  seed = runif(1, 0, .Machine$integer.max),
  bound_sims = FALSE,
  return_sims = FALSE,
  natural_course = TRUE,
  conf_level = 0.95,
  conf_type = "norm"
)
}
\arguments{
\item{gformula}{A g-formula object containing fitted models.}

\item{newdata}{A data.frame or matrix object containing new data to use for
simulation. Default is to use same dataset used to fit models. Useful for
generalizing or transporting results to a target population under
additional assumptions.}

\item{interventions}{A list of time-varying interventions}

\item{reference}{The name of the intervention to use as reference. Defaults
to the natural course.}

\item{n_samples}{Number of monte carlo samples to use to simulate histories.
Integer. Defaults to number of unique IDs in either original data set or
\code{newdata}.}

\item{n_boots}{Number of bootstrap samples to draw. Integer. Defaults to \code{0}.}

\item{start_time}{Start of follow up time for simulation. Integer or scalar.
Could be different than that specified during model fitting. Default is
time specified in \code{gformula} fit.}

\item{stop_time}{End of follow up time for simulation. Integer or scalar.
Could be different than that specified during model fitting. Default is
time specified in \code{gformula} fit.}

\item{seed}{Starting seed for random number generation.}

\item{bound_sims}{Should covariate simulations be restricted to the range of
the observed data? Boolean. Defaults to FALSE.}

\item{return_sims}{Flag for whether to return simulated covariate histories
for each sample. Boolean. Defaults to FALSE.}

\item{conf_level}{A scalar specifying the confidence level for bootstrapped
intervals in output.}

\item{conf_type}{A string specifying the type of bootstrap confidence
interval requested. One of "norm" or "perc".}
}
\description{
to approximate summation/integration over covariate history in g-formula
}
