% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{simulate_intervention}
\alias{simulate_intervention}
\title{Title}
\usage{
simulate_intervention(
  dt,
  intervention,
  id,
  time,
  start_time,
  stop_time,
  outcome_fit,
  covariate_fit,
  compevent_fit,
  censor_fit,
  lagged_histories,
  cumulative_histories,
  covariate_range,
  outcome_range,
  compevent_range,
  restrictions,
  bound_sims,
  return_sims,
  last_only,
  prediction
)
}
\arguments{
\item{prediction}{}
}
\description{
Title
}
