% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{gformula}
\alias{gformula}
\title{Define and fit (parametric) models for the g-formula}
\usage{
gformula(
  outcome_model,
  covariate_model,
  compevent_model = NULL,
  censor_model = NULL,
  visit_model = NULL,
  data,
  id,
  time,
  start_time = 0,
  stop_time = NULL,
  survival = FALSE,
  censor_compevent = TRUE,
  covs_baseline = NULL,
  covs_tv = NULL
)
}
\arguments{
\item{outcome_model}{Model specification for outcome (Y).}

\item{covariate_model}{Model specifications for time-varying covariates,
including potential treatment variables (L and A).}

\item{compevent_model}{Optional model specification for competing events (D).
Only implemented for survival models.}

\item{censor_model}{Optional model specification for censoring (C). Only used
to weight otherwise nonparametric estimates of the natural course such that
the estimand is the same as those targeted by the (parametric) models for
the outcome and covariates.}

\item{visit_model}{Optional model specification for visit process.}

\item{data}{A data.frame or matrix object containing the observed data.}

\item{id}{Name of unique id variable in \code{data}. Either a character vector or
something coercible to one.}

\item{time}{Name of follow up time variable in \code{data}. Either a character
vector or something coercible to one.}

\item{start_time}{Start of follow up time. Integer or scalar. Defaults to min
value of \code{time}.}

\item{stop_time}{End of follow up time. Integer or scalar. Defaults to max
value of \code{time}.}

\item{survival}{Outcome is a time-to-event/survival variable. Boolean.
Default is FALSE.}

\item{censor_compevent}{Flag for whether competing events should be treated
as censoring events (i.e. part of C) or modeled (i.e. part of L). Boolean.
Only implemented for survival models. Default is TRUE.}

\item{covs_baseline}{Names of time-fixed baseline covariates. Either a
character vector or something coercible to one.}

\item{covs_tv}{Names of time-varying covariates. Either a
character vector or something coercible to one.}
}
\value{
A gformula object with fitted models.
}
\description{
TBD
}
