% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{predict.gformula}
\alias{predict.gformula}
\title{Predict conditional expectations under time-varying interventions using the
g-formula}
\usage{
\method{predict}{gformula}(
  gformula,
  newdata = NULL,
  interventions = NULL,
  n_sims = NULL,
  n_boots = NULL,
  start_time = NULL,
  stop_time = NULL,
  seed = runif(1, 0, .Machine$integer.max),
  bound_sims = FALSE,
  return_sims = FALSE,
  natural_course = TRUE,
  last_only = TRUE,
  conf_level = 0.95,
  conf_type = "norm"
)
}
\arguments{
\item{conf_type}{}
}
\description{
TBD.
}
